@@ -12,33 +12,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.11",
-    "@angular/cdk": "^5.2.5",
-    "@angular/common": "^5.2.11",
-    "@angular/compiler": "^5.2.11",
-    "@angular/core": "^5.2.11",
-    "@angular/flex-layout": "5.0.0-beta.14",
-    "@angular/forms": "^5.2.11",
-    "@angular/http": "^5.2.11",
-    "@angular/material": "^5.2.5",
-    "@angular/material-moment-adapter": "^5.2.5",
-    "@angular/platform-browser": "^5.2.11",
-    "@angular/platform-browser-dynamic": "^5.2.11",
-    "@angular/router": "^5.2.11",
+    "@angular/animations": "6.0.4",
+    "@angular/cdk": "^6.2.1",
+    "@angular/common": "6.0.4",
+    "@angular/compiler": "6.0.4",
+    "@angular/core": "6.0.4",
+    "@angular/flex-layout": "6.0.0-beta.16",
+    "@angular/forms": "6.0.4",
+    "@angular/http": "6.0.4",
+    "@angular/material": "^6.2.1",
+    "@angular/material-moment-adapter": "^6.2.1",
+    "@angular/platform-browser": "6.0.4",
+    "@angular/platform-browser-dynamic": "6.0.4",
+    "@angular/router": "6.0.4",
     "@ngx-translate/core": "^9.1.1",
     "@types/openlayers": "4.6.9",
     "core-js": "^2.5.6",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
-    "moment": "^2.22.1",
+    "moment": "^2.22.2",
     "openlayers": "4.6.5",
-    "rxjs": "^5.5.10",
+    "rxjs": "^6.2.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.11",
-    "@angular/language-service": "^5.2.11",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "6.0.4",
+    "@angular/language-service": "6.0.4",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^6.0.111",
@@ -53,7 +53,8 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
-    "uglifyjs-webpack-plugin": "^1.2.5"
+    "typescript": "2.7.2",
+    "uglifyjs-webpack-plugin": "^1.2.5",
+    "@angular-devkit/build-angular": "~0.6.8"
   }
 }
